simplificacion de codigo
